Move alphAlignDir2perseidsAlignFile code
Can now be found at PerseusDL/alphAlignDir2perseidsAlignFile. The
commit history of this folder is retained in the new repository, for
historical referencing the original commits inside of this repository
stay where they are.

The new repository's name can still be discussed and changed if we come
up with a better solution.

Closes #35